updates readme, env handling and wasm lib

+++ update_wasm.sh
@@ -7,21 +7,23 @@ fn=${1##*/}
 mkdir temp
 unzip $fn -d temp
 cd temp
-if [[ "$OSTYPE" == "darwin"* ]]; then
-    sed -i '' "s|input = new URL('kdflib_bg.wasm', import.meta.url);|input = new URL('kdflib_bg.wasm', process.env.NEXT_PUBLIC_BASE_PATH);|" kdflib.js
-else
-    sed -i "s|input = new URL('kdflib_bg.wasm', import.meta.url);|input = new URL('kdflib_bg.wasm', process.env.NEXT_PUBLIC_BASE_PATH);|" kdflib.js
-fi
+# if [[ "$OSTYPE" == "darwin"* ]]; then
+#     sed -i '' "s|input = new URL('kdflib_bg.wasm', import.meta.url);|input = new URL('kdflib_bg.wasm', process.env.NEXT_PUBLIC_BASE_PATH);|" kdflib.js
+# else
+#     sed -i "s|input = new URL('kdflib_bg.wasm', import.meta.url);|input = new URL('kdflib_bg.wasm', process.env.NEXT_PUBLIC_BASE_PATH);|" kdflib.js
+# fi
 
 mv kdflib.js ../../js/kdflib.js
 mv kdflib.d.ts ../../js/kdflib.d.ts
 # mv kdflib.wasm.d.ts ../../js/kdflib.wasm.d.ts
-cp -r snippets/* ../../js/snippets/
+rsync -avh --delete snippets/ ../../js/snippets/
 
 basename=$(basename "$fn" .zip) # Remove .zip from filename
 temp=${basename#*_}
 version=${temp%-wasm} # Remove everything up to and including the first underscore
-rm ../../public/*.wasm
+if [ "$2" = "update_default" ]; then
+    rm ../../public/*.wasm
+fi
 
 mv kdflib_bg.wasm ../../public/kdflib_bg_$version.wasm
 
@@ -34,4 +36,13 @@ if grep -q "VITE_WASM_BIN=" .env; then
     sed -i.bak "s/VITE_WASM_BIN=.*/VITE_WASM_BIN=kdflib_bg_$version.wasm/" .env && rm .env.bak
 else
     echo "VITE_WASM_BIN=kdflib_bg_$version.wasm" >>.env
+fi
+
+# Update .env.example only if the script is called with argument "update_default"
+if [ "$2" = "update_default" ]; then
+    if grep -q "VITE_WASM_BIN=" .env.example; then
+        sed -i.bak "s/VITE_WASM_BIN=.*/VITE_WASM_BIN=kdflib_bg_$version.wasm/" .env.example && rm .env.example.bak
+    else
+        echo "VITE_WASM_BIN=kdflib_bg_$version.wasm" >>.env.example
+    fi
 fi
